Move amp helpers into amp folder lib
