+++ src/model/enhance-interactive-atom-elements.ts
@@ -1,28 +1,19 @@
-import { sanitiseHTML, clean } from '@root/src/model/clean';
+import { sanitiseHTML } from '@root/src/model/sanitise';
 
 // Some interactives contain HTML with unclosed tags etc. To
 // preserve (approximate) parity with Frontend we perform some basic
 // cleaning.
-export const enhance = (
-	elements: CAPIElement[],
-	isAmp: boolean = false,
-): CAPIElement[] => {
+export const enhance = (elements: CAPIElement[]): CAPIElement[] => {
 	return elements.map((element) => {
 		if (
 			element._type ===
 			'model.dotcomrendering.pageElements.InteractiveAtomBlockElement'
 		) {
-			if (isAmp) {
-				element.html = element.html
-					? clean(element.html)
-					: element.html;
-			} else {
-				element.html = element.html
-					? sanitiseHTML(element.html, {
-							ADD_TAGS: ['iframe', 'script'],
-					  })
-					: element.html;
-			}
+			element.html = element.html
+				? sanitiseHTML(element.html, {
+						ADD_TAGS: ['iframe', 'script'],
+				  })
+				: element.html;
 		}
 		return element;
 	});
